Initialize each cell with an empty array
Oops, needed to map each cell in the array to a new array, not just pass
it into the `.fill()` method. The old code was just creating one array
per row, so if a cell got "counted" in that row, every cell in that row
incremented as well! No wonder the code didn't work...