Working part 2
Not 100% sure if my assumption at the end to count coords "surrounded"
by the area is sound. Is it possible to have a coord on the "edge" of
the area? Then it'd still be part of the area, but not be surrounded.

Oh well, the answer works in this case.