Part one code, doesn't work
I know the bug: the problem is, as I'm looping through my plants, I'm
updating its state _in place_, which effects the state of the next plant
(if the previous one changed).

So instead, I need to loop through and compute the plants next state,
and store that _somewhere else_ (aka, not on the plant directly). Then,
loop through my plants again and set its state without recomputing it,
just grab it from the cached value.