Working part one and two
This one took me a while, and I'm still not happy with the solution. The
thing that took me the longest was figuring out I had to cache the
values for each of the gates. Similar to recursivly finding the
fibonacci sequence, you have to memoize things for the program to run
correctly.