Working day 11, all parts
I liked this one because of how simple I was able to make the code. That
is, I leveraged the `radix` param of `parseInt` and `toString` to do the
main heavy lifting for me. After that, I had some conversion function to
go back and forth between the states to make the whole thing work.

It was not as pretty as if I wrote my own custom base26 code (since I
have to go from "letters-only base26" -> "real base26" -> decimal
-> back to "real base26" -> and finally to "letters-only-base26"), but
hey, no one said it was the most optimized.