Still doesn't work, my logic needs to be more complex
I cheated here and looked at some solutions and the main trick is to
think of the grid as actual a three dimensional connected graph, where
the `tool` is the third dimension. Then, you use Dijkstra's algorithm
(although I don't really know how that works...) to find the (a)
shortest path to the target.