Working part two!
Reading the problem again, it isn't clear at all that the cave is a
"depth by depth" diszed grid. So, I just adding 50 worth of padding on
either side, looks to have done the trick.

The problem is, making a graph that is "depth by depth" or even
"target_y by target_y," my program ran for 30 minutes and didn't finish.
It is just too many nodes / edges.

This way, it is much more manageable for the computer.