This one does work, but I don't get it
I'm being stupid about the inner loop of permutations. I want my
perm_of_amounts array to be in increasing order, which amounts_range
already is in. Shouldn't I create my amounts first and permute that?
Or, just do the original combintations. Won't I get [0, 0, 0, 100] and
[100, 0, 0, 0] without having to permute them?

I'm confused, but I get the right answer... Need to debug this a bit to
see what is going on.