More not working code
I see what is wrong here, as soon as it finished the first call, it
stops. Instead, I need to generate a list of functions to call, and run
through that stack. When I get to the end, I won't add any more to the
stack.

I think that'll work...