After ~700,000,000 iterations, I still haven't found a solution
So, I googled it, and the answer actually totally makes sense. Rather
than trying to _grow_ our molecule from the `e` starting point, instead
start at the finished molecule and work backwards until to arrive at
`e`!

I found a program that did this, and from what I can, they start with
their molecule, then go through and do a single replacement with all
possible transforms (going backwards, that is). For instance, if a
normal growth is `B => TiRnFAr`, then find all instances of 'TiRnFAr'
and replace them with a B. Do this for every possible transformation
where there is a match.

There is a good chance the list will contain doubles around iteration 1
or 2 (think about it, Say I have `A => Ca` and `B => Ca` and starting
from `CaCaQ`, I run `A => Ca` to get `ACaQ` then on the second iteration
I run `B => Ca` to get `ABQ`, that is the same if I flipped those steps
[first getting `CaBQ` next `ABQ`]), so we'll want to `uniq` that list.
Additionally, the list we'll generate will be pretty huge, but some will
have been reduced quick than others. So, sort our list by size, and take
the smallest, say 1,000 molecules. That number I'm pretty sure is just a
guess, but we want it large enough that we don't accidentally throw away
the real solution. However, it probably doesn't matter, because again,
there are a lot of ways to grow the molecule (it seems) so in our case,
we're maximizing for the quickest reduction. No harm in doing that,
really.

So, that's it. You keep looping through all possible transformations,
uniq-ing that list, then taking the shortest 1000 and doing it again.
Eventually, we'll go from a single step over to 'e' and be done.