First stab at part two, doesn't work
I can't just take the number and add fifty, because that isn't right.
For instance, elf #2 will visit house 100 on his fiftith stop. Ah, it
just clicked: I need to take the elf number (the divisor) and multiply
that by fifty, and see if that is less than the house.