Working part two!
Had to completely change my approach here and do it analytically.
That is, a given layer has a "period" of oscillation of `(range - 1) * 2`.
If you think about it, if you have a layer of length 5, it starts at the
beginning. Next, it moves through the "bottom" 4 sections. Then (because
it is "starting" at the end) it moves the "top" 4 sections. So `(range -
1) * 2`!.

Next, if you want to see if you'll be "caught", take your delay _plus_
your depth and divide by the period. If the period divides that value,
you will be caught.

Again, this makes sense. In the first layer, there is no delay, so
you'll always be caught on the layers period.

If there is a layer of period 6 (range of 4), and it is at depth 3, then
you'll be caught if `(delay + 3) % 6 === 0`.

Given this, what I do is start with some delay, and begin testing them
against the layers. If `((delay + depth) % ((range - 1) * 2) === 0`,
then I _know_ I will get caught. So, I exit my loop, and tick up the
delay and try testing again.

This turns out to go _much_ faster than the simulation approach, namely
because as my delay increases, I have to recalculate the starting
position _from scratch_ each time. So if I have a delay of 1000, I have
to do 1000 ticks just to _start_ testing. Once that fails, I have to do
1001, and so on. When looking at the delay I'm testing, this number
quickly slows down and I spend all the time in the application "setting
up" the initial state to test.