Refactors code, supports granular dimension calculations
Couldn't figure this one out, so I took to the subreddit. The trick is,
each dimensions is periodic and _independent_, so you just calculate the
period of each dimension, then return the least-comon-multiple between
them, which is when the period will all be back in sync and we'll be
back at the starting position.