Test grid output
Going to have to debug this even more, I'm hoping that the ball will be
_at most_ one unit away from the paddle, and so we only have to decide
to move the paddle right when it asks for an input.

This is fairly clever, but tough! I feel like my setup for how `inputs`
and `outputs` work in my intcode computer isn't ideal for these
scenarios, but I'm not sure. I may end up refactoring that part.