Working part two!
Uggh, I just read the instructions wrong. I thought
that the "first in a direction" was looking
_specifically_ for the first *occupied* seat. Instead
I need to fine the first _seat_ in general, meaning
it can be empty *or* occupied. Dang, that took a
while to work through.

Along the way I was able to simplify some things as well,
remove some duplicate code.