Implements most of the logic, but is missing crucial piece
I need to look at the _unset_ immediate neighbors
of my active cubes. Basically the ones of the edges.

My first thought is to has some min / max tracking,
which is what I'm going to try and do (this ends
up being a bad idea / isn't going to work / increases
the complexity).