Begins recursive "orient" method on the pieces
The idea is, since I am merely setting "this piece
is the top left," I then say "OK, so by definition,
find the top left's right adjacent piece, and the
top left's bottom adjacent piece." I pass this in
with my `joins` arg.

I'm imagining I'll call it like

    piece.orientToConnections(RIGHT, BOTTOM)

And what that means is:

> Take the *first* connection (from our ordered connections)
> and orient both pieces so the connection matches with
> the `piece`s RIGHT side, and take the *second* connection
> and again orient both so the connection matches with
> the `piece`s BOTTOM side.

For this first step with two joins, each one will get me two
possibilities (The RIGHT join can be done with either piece flipped
vertically). Same for the BOTTOM join. If I take the _intersection_
of those two arrays, it should show me the orientation of
the current piece.

Then, based on that orientation, I can determine the orientation
of the RIGHT and BOTTOM pieces.

Then, I recursively build the orientations of those two pieces.
*If* the piece doesn't have a current orientation, then continue
with this recursion.

I also need a way to pass in this orientation. For the first call,
I don't have this, hence the intersection code. But for everything else,
I will have the piece's orientation, and so should be able to determine
the orientation of all its neighbors.