Removes cartesian call
This was the big slowdown in my flame graph. This can be improved
by looking at _one_ fixed orientation, and looping over all of
the other's orientations and all sides to test.

Basically, if two pieces fit, and I test each pair of orientations
and all sides, it is guaranteed that I'll get at least four hits,
if not eight. So, rather than waste time with that (which cartesian
is slow to start), I can just start iterating over the list of
orientations that I already have until I find a match.