Optimization complete!
I no longer create new objects for ops, modes, or values. Instead I
always update the same objects (modes / values are arrays) in place,
which helps a lot! Looks like around a 3x speed improvement.

Modes is a single array which is possible since we know the number of
params. We loop through params which means we don't hit the end of the
modes if the op doesn't need it.

Values needed to be separate arrays on each op since we spread this out
into the function call. Not a big deal since we create these arrays in
the constructor.