Bit of a hack optimization, but it works
So making a guess that _some bottom chunk_ of our paths can be discarded
to speed up the overall execution time really hits home this point: the
expensive part is making a frontier for each path.

If I wanted to redo this, an idea would be: make a frontier for each key
(plus the entrance) and add a "isReachable" check using our existing
keys. Each loop I'd just continue over the remaining keys, see which
ones are reachable, and make a new path from that.

That'd probably be WAY faster, but seems like a big re-write.