Fixed the issue
I can't just rely on the first index to see if this branches from the
same path, because I could skip the first one (as is often the case in
the 2nd step when the first connection is `start`). So instead just
reuse `dead_end` and only reset this once, it has the same effect.