Working!
Figured this out pretty quick. My intuition was correct in looking for a
"stable state." I did read the instructions incorrectly at first - do
1,000,000,000 *cycles* and not just 1b *revolutions* (so 4b
revolutions). No matter for the approach, but the trick was testing with
the sample input, and noticing that after I had seen a board position,
if I ran it through 1 more cycle, the board state was different. So its
periodic state isn't one. Once I was in the stable state, then I just
need to continuing running through until I determine the period.

I could have done this with another Set, but it was kind of nice to use
an array so I could see what the cycle was (A --> B --> C, etc). I did
double check that the first element in the array was indeed our match.

After that, just some math to determine how many more cycles I need to
run to get to 1 billion.