Adds some generatorics types
Used ChatGPT to convert the library to TS. The o1 model wasn't able to
handle it perfectly, but it was good enough to then plug into TS
Playground and generate a .d.ts file, which I copied here.

It seems like it is close enough to the real types, which is good enough
for me!